Fix and restructure IP tests
* Split up models and views tests into different files
* Update tests to match new auth system
* Default path argument in list_files to empty string
